requirements.txt
keras==2.13.1
tensorflow==2.13.0
scipy==1.10.1
scikit-image==0.20.0
Pillow==9.4.0
flask==2.2.2
numpy==1.24.3
gunicorn